renamed some folders to maintain the same structure
