+++ src/components/Board/assets/style.js
