+++ src/components/Cashier/assets/style.js
