+++ src/components/Chips/assets/style.js
